Undid some code moves which didnt need to be part of this PR.
